requirements.txt
numpy
yfinance
tradingview-ta
python-telegram-bot==13.12
